fisheries_matrices

initialize the transition, observation, and reward matrices given a transition function, reward function, and state space

fisheries_matrices(states = 0:20, actions = states,
  observed_states = states, reward_fn = function(x, a) pmin(x, a),
  f = ricker(1, 15), sigma_g = 0.1, sigma_m = 0.1,
  noise = c("lognormal", "uniform"))

Arguments

states

sequence of possible states

actions

sequence of possible actions

observed_states

sequence of possible observations

reward_fn

function of x and a that gives reward for tacking action a when state is x

f

transition function of state x and action a.

sigma_g

half-width of uniform shock or equivalent variance for log-normal

sigma_m

half-width of uniform shock or equivalent variance for log-normal

noise

distribution for noise, "lognormal" or "uniform"

Value

list of transitition matrix, observation matrix, and reward matrix

Details

assumes log-normally distributed observation errors and process errors